+++ frontend/source/class/agrammon/Application.js
@@ -16,10 +16,10 @@ qx.Class.define('agrammon.Application', {
         main: function() {
             this.base(arguments);
 
-            var rv = -1; // Return value assumes failure.
+            let rv = -1; // Return value assumes failure.
             if (navigator.appName == 'Microsoft Internet Explorer') {
-                var ua = navigator.userAgent;
-                var re  = new RegExp("MSIE ([0-9]{1,}[\.0-9]{0,})");
+                let ua = navigator.userAgent;
+                let re  = new RegExp("MSIE ([0-9]{1,}[\.0-9]{0,})");
                 if (re.exec(ua) != null) {
                     rv = parseFloat( RegExp.$1 );
                 }
@@ -37,24 +37,24 @@ qx.Class.define('agrammon.Application', {
                 qx.log.appender.Console;
             }
 
-                var param, params = this.__getParams();
-                for (var i=0; i<params.length; i++) {
+            let param, params = this.__getParams();
+            for (let i=0; i<params.length; i++) {
                 if (params[i] != null) {
                     param = params[i].split("=");
                 }
-                        if (param[0] == 'lang') {
+                if (param[0] == 'lang') {
                     qx.locale.Manager.getInstance().setLocale(param[1]);
-                        }
+                }
             }
 
-            var that = this;
+            let that = this;
 
             this.__rpc = agrammon.io.remote.Rpc.getInstance();
 
             qx.event.message.Bus.subscribe('agrammon.main.logout', this.__logout, this);
             qx.event.message.Bus.subscribe('agrammon.main.login',  this.__login,  this);
 
-            var root = this.getRoot();
+            let root = this.getRoot();
             root.setBlockerColor("#bfbfbf");
             root.setBlockerOpacity(0.5);
 
@@ -70,8 +70,9 @@ qx.Class.define('agrammon.Application', {
 
             let loginDialog =
                 new agrammon.module.user.Login(this.tr("Please authenticate yourself"));
+            this.__loginDialog = loginDialog;
 
-            for (var id in qx.core.Id.getInstance().getRegisteredObjects()) {
+            for (let id in qx.core.Id.getInstance().getRegisteredObjects()) {
                 this.debug('Id=', id);
             }
 
@@ -80,27 +81,27 @@ qx.Class.define('agrammon.Application', {
             }
 
             // the base layout of the page.
-            var main = new qx.ui.container.Composite(new qx.ui.layout.VBox());
+            let main = new qx.ui.container.Composite(new qx.ui.layout.VBox());
             main.set({ padding: 5 });
 
             root.add(main, { edge: 0 });
 
-            var title = new qx.ui.basic.Label().set({
+            let title = new qx.ui.basic.Label().set({
                 value          : 'AGRAMMON', // will be overwritten from config
                 font           : qx.bom.Font.fromString('14px bold sans-serif'),
                 textColor      : '#808080'
             });
 
-            var output     = new agrammon.module.output.Output(false);
-            var reference  = new agrammon.module.output.Output(true);
+            let output     = new agrammon.module.output.Output(false);
+            let reference  = new agrammon.module.output.Output(true);
 
-            var propEditor = new agrammon.module.input.PropTable('InputTable');
-            var navbar     = new agrammon.module.input.NavBar(propEditor);
+            let propEditor = new agrammon.module.input.PropTable('InputTable');
+            let navbar     = new agrammon.module.input.NavBar(propEditor);
 
-            var mainMenu;
+            let mainMenu;
 
-            var results;
-            var getCfgFunc = qx.lang.Function.bind(function(data, exc, id) {
+            let results;
+            let getCfgFunc = qx.lang.Function.bind(function(data, exc, id) {
                 if (exc == null) {
                     this.debug('getCfgFunc(): title='   +data.title.en);
                     this.debug('getCfgFunc(): version=' +data.version);
@@ -111,9 +112,9 @@ qx.Class.define('agrammon.Application', {
                     if (data.guiVariant != 'Regional') {
                         results = new agrammon.module.output.Results(output);
                     }
-                    var input      = new agrammon.module.input.Input(propEditor, navbar, results);
-                    var tabview    = new agrammon.module.Main(input, output, reference);
-                    var editMenu   = new agrammon.ui.menu.NavMenu(navbar);
+                    let input      = new agrammon.module.input.Input(propEditor, navbar, results);
+                    let tabview    = new agrammon.module.Main(input, output, reference);
+                    let editMenu   = new agrammon.ui.menu.NavMenu(navbar);
                     mainMenu       = new agrammon.ui.menu.MainMenu(tabview, title, editMenu);
 
                     main.add(mainMenu);
@@ -141,11 +142,12 @@ qx.Class.define('agrammon.Application', {
             this.__rpc.callAsync( getCfgFunc, 'get_cfg');
 
             this.__authenticate = function(data, exc, id) {
+                // console.log('Application.__authenticate():', data, exc, id);
                 if (exc == null) {
-                    var username  = data.username;
-                    var role      = data.role;
-                    var news      = data.news;
-                    var lastLogin = String(data.lastLogin);
+                    let username  = data.username;
+                    let role      = data.role;
+                    let news      = data.news;
+                    let lastLogin = String(data.lastLogin);
                     qx.event.message.Bus.dispatchByName(
                         'agrammon.info.setUser',
                         { username : username, role : role }
@@ -194,30 +196,23 @@ qx.Class.define('agrammon.Application', {
         __authenticate: null,
         __retry:        null,
         __rpc:          null,
-
-        __supports_html5_storage: function() {
-            try {
-                return 'localStorage' in window && window['localStorage'] !== null;
-            } catch (e) {
-                return false;
-            }
-        },
+        __loginDialog:  null,
 
         close : function(e) {
             this.base(arguments);
-            this.debug('Application.close()');
+            console.log('Application.close()');
             // Prompt user
             // return "AGRAMMON: Do you really want to close the application?";
         },
 
         terminate : function(e) {
             this.base(arguments);
-            this.debug('Application.terminate()');
+            console.log('Application.terminate()');
         },
 
         __getParams : function() {
-            var params = "";
-            var urlParams = window.location.search;
+            let params = "";
+            let urlParams = window.location.search;
             if (urlParams.length > 0) {
                 urlParams = urlParams.substr(1, urlParams.length);
                 if (params != null) {
@@ -228,14 +223,9 @@ qx.Class.define('agrammon.Application', {
         },
 
         __login: function(msg) {
-            var userData     = msg.getData();
+            let userData     = msg.getData();
             this.__retry = userData.retry;
             this.debug('__login(' + userData.username + ')');
-            if (this.__supports_html5_storage() && userData.remember) {
-                localStorage.setItem('agrammonUsername', userData.username);
-                localStorage.setItem('agrammonPassword', userData.password);
-                localStorage.setItem('agrammonRemember', userData.remember);
-            }
             if (userData.sudoUsername === undefined) {
                 userData.sudoUsername = null;
             }
@@ -243,16 +233,18 @@ qx.Class.define('agrammon.Application', {
         },
 
         __logout: function() {
+            // console.log('Application.__logout()');
             this.__rpc.callAsync( qx.lang.Function.bind(this.__logoutFunc,this), 'logout');
             qx.event.message.Bus.dispatchByName('agrammon.NavBar.clearTree', null);
             qx.event.message.Bus.dispatchByName('agrammon.input.select');
         },
 
         __logoutFunc: function(data, exc, id) {
+            // console.log('Application.__logoutFunc():', data, exc, id);
             if (exc == null || exc == 403) {
-                if (data.sudoUser) {
-                    var infoOnly = true;
-                    var dialog = new agrammon.ui.dialog.Confirm(
+                if (data && data.sudoUser) {
+                    let infoOnly = true;
+                    let dialog = new agrammon.ui.dialog.Confirm(
                         this.tr("End change user"),
                         this.tr("Returning from %1 to %2", data.sudoUser, data.username),
                         qx.lang.Function.bind(function() {
@@ -270,7 +262,7 @@ qx.Class.define('agrammon.Application', {
             else {
                 alert(exc);
             }
-            new agrammon.module.user.Login(this.tr("Please authenticate yourself")).open();
+            this.__loginDialog.open();
         }
     }
 });