+++ runWeb.sh
@@ -1,2 +1,3 @@
 #! /bin/bash
-raku -Ilib bin/agrammon.pl6 --cfg-file=etc/agrammon.single.yaml web version6/End.nhd
+export PERL5LIB=thirdparty/lib/perl5
+raku -Ilib bin/agrammon.raku --cfg-file=etc/agrammon.single.yaml web version6/End.nhd